Moves reference to {this.state.text} down
First code snippet that included {this.state.text} did not have an explanation for it in the following paragraph, and it was absent from the snippet that included an explanation in the paragraph. This change would move {this.state.text} to the area where it is explained to avoid confusion while following the walkthrough.